[CONTROLLER/COMMON] refactors common dir
