+++ server/controller/common/net.go
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2023 Yunshan Networks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package common
+
+import (
+	"fmt"
+	"net"
+	"time"
+)
+
+func IsTCPActive(ip string, port int) error {
+	conn, err := net.DialTimeout("tcp", fmt.Sprintf("%s:%d", ip, port), 2*time.Second)
+	if err != nil {
+		return err
+	} else {
+		if conn != nil {
+			conn.Close()
+		} else {
+			return fmt.Errorf("check tcp alive failed (ip:%s, port:%d)", ip, port)
+		}
+	}
+	return nil
+}
