+++ server/controller/http/service/vtap_group_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package common
+package service
 
 import "testing"
