+++ server/controller/trisolaris/node/node.go
@@ -275,7 +275,7 @@ func (n *NodeInfo) generateControllerInfo() {
 	if err == nil {
 		localAZs := make([]string, 0, len(localConns))
 		for _, localConn := range localConns {
-			if IsValueInSliceString(localConn.AZ, localAZs) == false {
+			if Contains(localAZs, localConn.AZ) == false {
 				localAZs = append(localAZs, localConn.AZ)
 			}
 		}
@@ -713,7 +713,7 @@ func (n *NodeInfo) generatePlatformData() {
 		if len(localAZs) == 0 {
 			n.updatePlatformData(allCompletePlatformDataExceptPod)
 		} else {
-			if IsValueInSliceString(CONN_DEFAULT_AZ, localAZs) {
+			if Contains(localAZs, CONN_DEFAULT_AZ) {
 				regionToPlatformDataOnlyPod := n.metaData.GetPlatformDataOP().GetRegionToPlatformDataOnlyPod()
 				if regionPlatformData, ok := regionToPlatformDataOnlyPod[localRegion]; ok {
 					platformData := metadata.NewPlatformData("platformData", "", 0, PLATFORM_DATA_FOR_INGESTER_2)
